# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /viewings/{token}:
    get:
      operationId: getViewingBooking
      tags: [viewings]
      summary: Get a booked viewing from its customer link
      description: Resolves the secure one-off link sent to the customer so they can manage their own viewing.
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: The booked viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewingBooking"
        "404":
          description: Link is not valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Unable to load the viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /viewings/{token}/reschedule:
    post:
      operationId: rescheduleViewing
      tags: [viewings]
      summary: Move a booked viewing to another slot
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewingReschedule"
      responses:
        "200":
          description: The rescheduled viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewingBooking"
        "400":
          description: Requested time is not bookable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Link is not valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Requested slot has just been taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Unable to reschedule the viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /viewings/{token}/cancel:
    post:
      operationId: cancelViewing
      tags: [viewings]
      summary: Cancel a booked viewing
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewingCancellation"
      responses:
        "200":
          description: The cancelled viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewingBooking"
        "400":
          description: Cancellation failed validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Link is not valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Unable to cancel the viewing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
          minLength: 1
      required:
        - error
...
    ViewingBooking:
      type: object
      additionalProperties: false
      properties:
        reference:
          type: string
          minLength: 1
        status:
          type: string
          enum: [pending, booked, cancelled]
        customerName:
          type: string
        appointmentAt:
          type: [string, "null"]
          format: date-time
        cancelledAt:
          type: [string, "null"]
          format: date-time
        durationMinutes:
          type: number
        timezone:
          type: string
        vehicleTitle:
          type: [string, "null"]
        vehicleUrl:
          type: [string, "null"]
        calendarIcs:
          type: [string, "null"]
        canChange:
          description: False once the viewing is cancelled or its start time has passed.
          type: boolean
      required:
        - reference
        - status
        - customerName
        - appointmentAt
        - cancelledAt
        - timezone
        - vehicleTitle
        - vehicleUrl
        - calendarIcs
        - canChange
    ViewingReschedule:
      type: object
      additionalProperties: false
      properties:
        appointmentAt:
          type: string
          format: date-time
      required:
        - appointmentAt
    ViewingCancellation:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: [string, "null"]
          maxLength: 500
...